requirements.txt
cupy-cuda12x
opencv-python
numpy
requests
pillow